Move `Lirical` to `lirical-core`.
