+++ lirical-beta/src/main/java/org/monarchinitiative/lirical/beta/cmd/PrioritizeWithSquirlsCommand.java
@@ -1,6 +1,6 @@
 package org.monarchinitiative.lirical.beta.cmd;
 
-import org.monarchinitiative.lirical.configuration.Lirical;
+import org.monarchinitiative.lirical.core.Lirical;
 import org.monarchinitiative.lirical.core.analysis.AnalysisData;
 import org.monarchinitiative.lirical.core.analysis.LiricalParseException;
 import org.monarchinitiative.lirical.core.model.GenesAndGenotypes;
