+++ lirical-core/src/main/java/org/monarchinitiative/lirical/core/io/VariantParserFactory.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.lirical.io;
+package org.monarchinitiative.lirical.core.io;
 
 import org.monarchinitiative.lirical.core.model.GenomeBuild;
 import org.monarchinitiative.svart.assembly.GenomicAssembly;
