+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/analysis/BaseAnalysisDataParser.java
@@ -3,8 +3,8 @@ package org.monarchinitiative.lirical.io.analysis;
 import org.monarchinitiative.lirical.core.analysis.AnalysisDataParser;
 import org.monarchinitiative.lirical.core.analysis.LiricalParseException;
 import org.monarchinitiative.lirical.core.model.*;
-import org.monarchinitiative.lirical.io.VariantParser;
-import org.monarchinitiative.lirical.io.VariantParserFactory;
+import org.monarchinitiative.lirical.core.io.VariantParser;
+import org.monarchinitiative.lirical.core.io.VariantParserFactory;
 import org.monarchinitiative.phenol.annotations.formats.GeneIdentifier;
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoAssociationData;
 import org.slf4j.Logger;
